Add files via upload
In this questionnaire, a new sort order was developed to provide further access to the questions for a Horizon2020-DMP. While the Horizon2020 template focuses on the FAIR data principles, the sorting in this questionnaire is intended to illustrate content-related relationships and thus facilitate the answering of the questions. In the comment field next to each question you will find an indication of which original H2020 question the current question in RDMO is aimed at. The numbering does not exist in the H2020 template, it was added for a better overview.